create teacher dto class and pass it through endpoints

+++ src/main/java/com/jj/Gradebook/service/teacher/TeacherServiceImpl.java
@@ -1,6 +1,7 @@
 package com.jj.Gradebook.service.teacher;
 
 import com.jj.Gradebook.dao.TeacherRepository;
+import com.jj.Gradebook.dto.TeacherDTO;
 import com.jj.Gradebook.entity.Teacher;
 import com.jj.Gradebook.exceptions.EntityAlreadyExistException;
 import com.jj.Gradebook.exceptions.EntityNotFoundException;
@@ -8,6 +9,8 @@ import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,16 +21,42 @@ public class TeacherServiceImpl implements TeacherService{
     private final TeacherRepository teacherRepository;
 
     @Override
-    public List<Teacher> findAll() {
-        return teacherRepository.findAll();
+    public List<TeacherDTO> findAll() {
+        List<TeacherDTO> result = new ArrayList<>();
+        List<Teacher> data = teacherRepository.findAll();
+
+        for (Teacher teacher: data){
+            result.add(
+                    new TeacherDTO(
+                            teacher.getTeacherId(),
+                            teacher.getFirstName(),
+                            teacher.getLastName(),
+                            teacher.getUser().getPesel(),
+                            teacher.getUser().getEmail(),
+                            new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfBirth().getTime()),
+                            new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfEmployment().getTime())
+                    )
+            );
+        }
+
+        return result;
     }
 
     @Override
-    public Teacher findById(int id) throws EntityNotFoundException {
+    public TeacherDTO findById(int id) throws EntityNotFoundException {
         Optional<Teacher> result = teacherRepository.findById(id);
 
         if(result.isPresent()){
-            return  result.get();
+            Teacher teacher = result.get();
+            return new TeacherDTO(
+                    teacher.getTeacherId(),
+                    teacher.getFirstName(),
+                    teacher.getLastName(),
+                    teacher.getUser().getPesel(),
+                    teacher.getUser().getEmail(),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfBirth().getTime()),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfEmployment().getTime())
+            );
         }
         else{
             throw new EntityNotFoundException("No teacher with id - " + id);
@@ -36,10 +65,19 @@ public class TeacherServiceImpl implements TeacherService{
     }
 
     @Override
-    public Teacher findByPesel(String pesel) throws EntityNotFoundException {
+    public TeacherDTO findByPesel(String pesel) throws EntityNotFoundException {
         Optional<Teacher> result = Optional.ofNullable(teacherRepository.findTeacherByUserPesel(pesel));
         if(result.isPresent()){
-            return result.get();
+            Teacher teacher = result.get();
+            return new TeacherDTO(
+                    teacher.getTeacherId(),
+                    teacher.getFirstName(),
+                    teacher.getLastName(),
+                    teacher.getUser().getPesel(),
+                    teacher.getUser().getEmail(),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfBirth().getTime()),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfEmployment().getTime())
+            );
         }
         else {
             throw new EntityNotFoundException("No teacher with pesel - " + pesel);
@@ -49,21 +87,36 @@ public class TeacherServiceImpl implements TeacherService{
 
     @Override
     @Transactional
-    public Teacher save(Teacher teacher) throws EntityAlreadyExistException {
+    public TeacherDTO save(Teacher teacher) throws EntityAlreadyExistException {
         Optional<Teacher> existingTeacher = Optional.ofNullable(teacherRepository.findTeacherByUserPesel(teacher.getUser().getPesel()));
 
         if (existingTeacher.isEmpty()){
-            return teacherRepository.save(teacher);
+            Teacher savedTeacher = teacherRepository.save(teacher);
+            return new TeacherDTO(
+                    teacher.getTeacherId(),
+                    teacher.getFirstName(),
+                    teacher.getLastName(),
+                    teacher.getUser().getPesel(),
+                    teacher.getUser().getEmail(),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfBirth().getTime()),
+                    new SimpleDateFormat("dd.MM.yyyy").format(teacher.getDateOfEmployment().getTime())
+            );
         }
         else {
-            throw new EntityAlreadyExistException("Teacher already exist");
+            throw new EntityAlreadyExistException("Teacher with this pesel already exists");
         }
 
     }
 
     @Override
     @Transactional
-    public void deleteById(int id) {
-        teacherRepository.deleteById(id);
+    public void deleteById(int id) throws EntityNotFoundException {
+        Optional<Teacher> existingTeacher = teacherRepository.findById(id);
+        if (existingTeacher.isPresent()){
+            teacherRepository.deleteById(id);
+        }
+        else {
+            throw new EntityNotFoundException("No teacher with id - " + id);
+        }
     }
 }
